Fix build img

+++ build_kata_img.sh
@@ -9,7 +9,9 @@ readonly script_dir="$(dirname $(readlink -f $0))"
 echo "Will create some new image from $img in $rootfsDir"
 
 # Create /sbin/init (why does not exist?)
-ln -s /usr/lib/systemd/systemd ${rootfsDir}/sbin/init
+[ -e "${rootfsDir}/sbin/init" ] || ln -s /usr/lib/systemd/systemd ${rootfsDir}/sbin/init
+# Create /lib/systemd/systemd (why does not exist?)
+[ -e "${rootfsDir}/lib/systemd/systemd" ] || (mkdir -p ${rootfsDir}/lib/systemd/; ln -s /usr/lib/systemd/systemd ${rootfsDir}/lib/systemd/systemd)
 
 ${script_dir}/osbuilder/image-builder/image_builder.sh -o "$img" "$rootfsDir"
